nestech_telegram_project/requirements.txt
python-telegram-bot>=21.4
python-dotenv
